calendar-sync/requirements.txt
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
squareup==34.0.0
python-dotenv==1.0.1
Flask==3.0.2
